+++ python/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Option to disable processing of "style" tags. [#45](https://github.com/Stranger6667/css-inline/issues/45)
+
 ## [0.3.2] - 2020-07-09
 
 ### Fixed